break out pixel-to-bitstring conversion
